Addressed a concern regarding consent
A number of readers have reacted strongly to a deliberately provocative set of scenes in this chapter. Among the criticisms, an issue surrounding consent and power dynamics between Ferd and Vemik was raised, which we believe to be both valid and in-line with what we intended to communicate. Therefore:

* Vemik more explicitly shows concern for Ferd's vulnerability
* narratively tied in with his existing relations with Tilly, where of course the power difference is vastly larger

This will not address everyone's concerns, but that is by design; we intend to discomfit a bit around alien psychology and morality, because good fiction should challenge the readers. The above more closely aligns with our intent and we feel it is an appropriate correction.